I fixed a minor problem (arising from the use of a more recent C++ compiler) with the std::abs function being passed an 'unsigned int' type.

+++ monophyllon/spacetime_geometry.cpp
@@ -345,7 +345,7 @@ double Spacetime::chorogenesis()
         derror += d;
       }
       else {
-        derror += std::abs(d - dpopulation[i]);
+        derror += (unsigned) std::abs(int(d - dpopulation[i]));
       }
     }
 #ifdef VERBOSE
